make PoolBlockingPeriod to netcoreapp specific
