+++ src/System.Data.SqlClient/src/System/Data/SqlClient/PoolBlockingPeriod.NetCoreApp.cs
